requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0
numpy>=1.20.0
pandas>=1.3.0
ortools>=9.0.0 